added harvest button

+++ lib/components/plants/crop.dart
@@ -3,9 +3,9 @@ import 'plant.dart';
 class Crop extends Plant {
   Crop({
     required String name,
-    required int growthTime,
-    required int sellPrice,
-    required int incomeRate,
+    required double growthTime,
+    required double sellPrice,
+    required double incomeRate,
     required int rarity,
     required String imagePath,
     required String spritePath,
